Use event_name to set concurrency group
There was a problem with creating a release from the github web
interface: the tag and the release are created at about the same moment.
This could lead to the release build being cancelled, which would
prevent the pypi artifacts from being uploaded since they would NOT
be uploaded for the build of the tag.

This speculatively fixes the problem.